requirements.txt
selenium>=4.15.2
python-dotenv>=1.0.0
schedule>=1.2.1
google-generativeai>=0.3.2
webdriver_manager>=4.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
python-linkedin-v2>=0.8.0 